Membership_Database_and_API/requirements.txt
Django == 5.1.6
djangorestframework == 3.15.2
python-dotenv == 1.0.1
mysqlclient == 2.2.7
pillow==11.1.0
django-filter==25.1
